requirements.txt
flask
python-mpd2
youtube-dl
youtube-search-python
music-tag
pydub
